Code changes to optimised code and incorporate css classes and dynemic content.
