+++ tree/master/cloud/src/solution/Microsoft.Legal.MatterCenter.Web/wwwroot/css/viewspecific/CreateMatter.css
@@ -1589,6 +1589,24 @@ left: -19px;
        left: -35px;
     top: 11px;
 }
+
+.directiveMatterExtraFields{
+    padding-right:190px;
+    float:right;
+}
+
+.directiveMatterExtraBoolField{
+    padding-left:60px;
+    padding-bottom:13px;
+    padding-top:8px;
+    float:left;
+}
+
+.directiveMatterExtraBoolLabel{
+    padding-top:6px;
+    float:left;
+}
+
 #SettingsPermissions .dropdown-menu li {
         max-width: 295px;
     min-width: 295px;
